Refactoring Reducer Logic Using Functional Decomposition and Reducer Composition https://redux.js.org/recipes/structuring-reducers/refactoring-reducer-example
