+++ src/ducks/soknad/reducers.js
@@ -1,9 +1,7 @@
 import { STATUS } from '../../services/utils';
 
 import * as Types from './types';
-
-import { strengTilInt } from '../../utils/streng';
-import { formatterDatoTilISO } from '../../utils/dato';
+import { oppdaterOK, oppdaterPeriode, oppdaterSoknad } from './transforms';
 
 /**
  * Reducers
@@ -25,130 +23,16 @@ export default function reducer(state = initialState, action) {
     case Types.FEILET:
       return { ...state, status: STATUS.ERROR, data: action.data };
     case Types.OK: {
-      const soknadData = action.data;
-      const { behandlingID, soeknadDokument } = soknadData;
-
-      return {
-        ...state,
-        status: STATUS.OK,
-        data: {
-          behandlingID,
-          soeknadDokument,
-        },
-      };
+      return oppdaterOK(state, action);
     }
     case Types.RESET:
       return { ...initialState };
+
     case Types.OPPDATER_PERIODE: {
-      const { oppholdsPeriode } = action.data;
-
-      const soknad = {
-        ...state.data.soeknadDokument,
-        oppholdUtland: {
-          ...state.data.soeknadDokument.oppholdUtland,
-          oppholdsPeriode: {
-            fom: oppholdsPeriode.fom,
-            tom: oppholdsPeriode.tom,
-          },
-        },
-      };
-
-      return { ...state, data: { ...state.data, soeknadDokument: soknad } };
+      return oppdaterPeriode(state, action);
     }
     case Types.OPPDATER_SOKNAD: {
-      const { dokument } = action;
-
-      const soknad = {
-        ...state.data.soeknadDokument,
-        arbeidsinntekt: {
-          ...state.data.soeknadDokument.arbeidsinntekt,
-          inntektNorskIPerioden: strengTilInt(dokument.inntektNorskIPerioden),
-          inntektUtenlandskIPerioden: strengTilInt(dokument.inntektUtenlandskIPerioden),
-          inntektNaeringIPerioden: strengTilInt(dokument.inntektNaeringIPerioden),
-        },
-        arbeidNorge: {
-          ...state.data.soeknadDokument.arbeidNorge,
-          fullmektigFirma: dokument.fullmektigFirma,
-          fullmektigGateadresse: dokument.fullmektigGateadresse,
-          fullmektigPostnr: dokument.fullmektigPostnr,
-          fullmektigPoststed: dokument.fullmektigPoststed,
-          fullmektigRegion: dokument.fullmektigRegion,
-          fullmektigLandKode: dokument.fullmektigLand,
-        },
-        arbeidUtland: dokument.arbeidUtland,
-        juridiskArbeidsgiverNorge: {
-          antallAnsatte: dokument.antallAnsatte,
-          utsendteNeste12Mnd: dokument.utsendteNeste12Mnd,
-          antallAdmAnsatte: dokument.antallAdmAnsatte,
-          andelOmsetningINorge: dokument.andelOmsetningINorge,
-          andelOppdragINorge: dokument.andelOppdragINorge,
-          andelKontrakterINorge: dokument.andelKontrakterINorge,
-          arbeidstakereRekruttertILand: dokument.arbeidstakereRekruttertILand || null,
-          oppdragsKontrakterIHovedsakInngaattILand: dokument.oppdragsKontrakterIHovedsakInngaattILand || null,
-          ekstraArbeidsgivere: dokument.ekstraArbeidsgivere || [],
-        },
-        arbeidsgiversBekreftelse: {
-          ...state.data.soeknadDokument.arbeidsgiversBekreftelse,
-          arbeidsgiverBekrefterUtsendelse: dokument.arbeidsgiverBekrefterUtsendelse,
-          arbeidstakerAnsattUnderUtsendelsen: dokument.arbeidstakerAnsattUnderUtsendelsen,
-          erstatterArbeidstakerenUtsendte: dokument.erstatterArbeidstakerenUtsendte,
-          arbeidstakerTidligereUtsendt24Mnd: dokument.arbeidstakerTidligereUtsendt24Mnd,
-          arbeidsgiverBetalerArbeidsgiveravgift: dokument.arbeidsgiverBetalerArbeidsgiveravgift,
-          trygdeavgiftTrukketGjennomSkatt: dokument.trygdeavgiftTrukketGjennomSkatt,
-          trygdeavgiftTrukketGjennomSkattDato: dokument.trygdeavgiftTrukketGjennomSkattDato ? formatterDatoTilISO(dokument.trygdeavgiftTrukketGjennomSkattDato) : null,
-        },
-        oppholdUtland: {
-          ...state.data.soeknadDokument.oppholdUtland,
-          oppholdsPeriode: {
-            fom: formatterDatoTilISO(dokument.oppholdUtlandFom),
-            tom: formatterDatoTilISO(dokument.oppholdUtlandTom),
-          },
-          oppholdslandKoder: dokument.oppholdsland,
-          sammeAdresseSomArbeidsgiver: dokument.sammeAdresseSomArbeidsgiver,
-          ektefelleEllerBarnINorge: null,
-          forutgaendeBostedINorge: dokument.forutgaendeBostedINorge,
-          studentSemester: null,
-          studentFinansieringKode: null,
-        },
-        foretakUtland: dokument.foretakUtland.filter(foretakUtland => (
-          foretakUtland.navn
-          && foretakUtland.orgnr
-          && foretakUtland.adresse
-        )),
-        bosted: {
-          intensjonOmRetur: null,
-          familiesBostedLandKode: dokument.familiesBosted,
-          antallMaanederINorge: null,
-          EOSBarnetrygdFraNAV: dokument.EOSBarnetrygdFraNAV,
-          adresseIUtlandet: dokument.adresseIUtlandet,
-          oppgittAdresse: {
-            gatenavn: dokument.oppgittAdresseGatenavn,
-            husnummer: dokument.oppgittAdresseHusnummer,
-            region: dokument.oppgittAdresseRegion,
-            postnummer: dokument.oppgittAdressePostnummer,
-            poststed: dokument.oppgittAdressePoststed,
-            landKode: dokument.oppgittAdresseLand,
-          },
-        },
-        maritimtArbeid: dokument.maritimtArbeid.map(maritimtArbeid => ({
-          navn: maritimtArbeid.navn ? maritimtArbeid.navn : null,
-          fartsomradeKode: maritimtArbeid.fartsomradeKode ? maritimtArbeid.fartsomradeKode : null,
-          flaggLandKode: maritimtArbeid.flaggLandKode ? maritimtArbeid.flaggLandKode : null,
-          installasjonsLandKode: maritimtArbeid.installasjonsLandKode ? maritimtArbeid.installasjonsLandKode : null,
-          territorialfarvann: maritimtArbeid.territorialfarvann ? maritimtArbeid.territorialfarvann : null,
-        })),
-        selvstendigArbeid: {
-          erSelvstendig: dokument.erSelvstendig,
-          selvstendigForetak: dokument.selvstendigForetak,
-        },
-        personOpplysninger: {
-          utenlandskIdent: dokument.utenlandskIdent,
-          medfolgendeFamilie: [],
-          medfolgendeAndre: null,
-        },
-      };
-
-      return { ...state, data: { ...state.data, soeknadDokument: soknad } };
+      return oppdaterSoknad(state, action);
     }
     default:
       return state;
